fix long type convert

+++ main.cpp
@@ -4211,7 +4211,7 @@ GValue phpgtk_get_gvalue(Php::Value phpgtk_value, GType type_column)
         case G_TYPE_LONG:
         {
             // Cast
-            int b = (int64_t)phpgtk_value;
+            long b = (long)phpgtk_value;
 
             g_value_init(&gtk_value, G_TYPE_LONG);
             g_value_set_long(&gtk_value, b);
